requirements.txt
streamlit
pandas
numpy
transformers
torch
sentence-transformers
chromadb
faiss-cpu
requests
beautifulsoup4
scikit-learn
tqdm
pytest